@@ -1,1 +1,1 @@
-ruby-2.3.0
+2.6.6
