@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module BlueJay
-  VERSION = '4.5.8'
+  VERSION = '4.6.0'
 end
